+++ packages/unmock-core/src/console/types.ts
@@ -1,0 +1,8 @@
+export type LogMessage = string;
+
+export interface ILogEntry {
+  message: LogMessage;
+  type: LogType;
+}
+
+export type LogType = "error" | "instruct";