+++ packages/unmock-node/src/utils.ts
@@ -1,32 +1,42 @@
+import appRootPath from "app-root-path";
 import debug from "debug";
 import fs from "fs";
+import { uniq } from "lodash";
 import path from "path";
 
 const debugLog = debug("unmock:node:utils");
 
-const DEFAULT_SERVICE_SUBDIRECTORY = "__unmock__";
+export const resolveUnmockRootDirectory = (): string =>
+  appRootPath.resolve("__unmock__");
 
 /**
- * Resolve the absolute path to the directory where services live, with the flow:
- * 1. Injected directory
- * 2. Environment variable
- * 3. ${process.cwd()}/__unmock__
+ * Resolve the absolute paths to unmock directories.
+ * Only checks that the directories exist.
+ * The returned list may be empty.
  */
-// TODO: Use `app-root-path` instead of `process.cwd()`?
-export const resolveServicesDirectory = (servicesDirOpt?: string) => {
-  const servicesDirectory = path.resolve(
-    servicesDirOpt ||
-      process.env.UNMOCK_SERVICES_DIRECTORY ||
-      path.join(process.cwd(), DEFAULT_SERVICE_SUBDIRECTORY),
+export const resolveUnmockDirectories = (): string[] => {
+  const defaultUnmockDirectories = [
+    resolveUnmockRootDirectory(),
+    appRootPath.resolve("node_modules/@unmock"),
+    // Also search process cwd, helpful with `npm link` for example
+    // where app-root-path fails due to a symlink
+    // without using APP_ROOT_PATH=`pwd`
+    path.resolve(`${process.cwd()}/__unmock__`),
+    path.resolve(`${process.cwd()}/node_modules/@unmock`),
+  ];
+
+  const triedDirectories = process.env.UNMOCK_SERVICES_DIRECTORY
+    ? defaultUnmockDirectories.concat(
+        path.resolve(process.env.UNMOCK_SERVICES_DIRECTORY),
+      )
+    : defaultUnmockDirectories;
+
+  const uniqueTriedDirectories = uniq(triedDirectories);
+
+  debugLog(`Checking for directories: ${triedDirectories}`);
+
+  return uniqueTriedDirectories.filter(
+    (directory: string) =>
+      fs.existsSync(directory) && fs.statSync(directory).isDirectory(),
   );
-
-  debugLog(`Resolved services directory: ${servicesDirectory}`);
-  if (
-    fs.existsSync(servicesDirectory) &&
-    fs.statSync(servicesDirectory).isDirectory()
-  ) {
-    return servicesDirectory;
-  }
-
-  throw new Error(`Directory ${servicesDirectory} does not exist`);
 };